pyproject.toml
[tool.poetry]
name = "radio-epg"
version = "0.1.0"
description = "Radio EPG Updater"
authors = ["Roy B. <royutil@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.7"
pytz = "^2019.3"
attrs = "^19.3.0"
"aws-cdk.assets" = "^1.18.0"
"aws-cdk.aws-apigateway" = "^1.18.0"
"aws-cdk.aws-applicationautoscaling" = "^1.18.0"
"aws-cdk.aws-autoscaling" = "^1.18.0"
"aws-cdk.aws-autoscaling-common" = "^1.18.0"
"aws-cdk.aws-autoscaling-hooktargets" = "^1.18.0"
"aws-cdk.aws-certificatemanager" = "^1.18.0"
"aws-cdk.aws-cloudformation" = "^1.18.0"
"aws-cdk.aws-cloudfront" = "^1.18.0"
"aws-cdk.aws-cloudwatch" = "^1.18.0"
"aws-cdk.aws-cloudwatch-actions" = "^1.18.0"
"aws-cdk.aws-codebuild" = "^1.18.0"
"aws-cdk.aws-codecommit" = "^1.18.0"
"aws-cdk.aws-codepipeline" = "^1.18.0"
"aws-cdk.aws-ec2" = "^1.18.0"
"aws-cdk.aws-ecr" = "^1.18.0"
"aws-cdk.aws-ecr-assets" = "^1.18.0"
"aws-cdk.aws-ecs" = "^1.18.0"
"aws-cdk.aws-elasticloadbalancing" = "^1.18.0"
"aws-cdk.aws-elasticloadbalancingv2" = "^1.18.0"
"aws-cdk.aws-events" = "^1.18.0"
"aws-cdk.aws-events-targets" = "^1.18.0"
"aws-cdk.aws-iam" = "^1.18.0"
"aws-cdk.aws-kms" = "^1.18.0"
"aws-cdk.aws-lambda" = "^1.18.0"
"aws-cdk.aws-logs" = "^1.18.0"
"aws-cdk.aws-route53" = "^1.18.0"
"aws-cdk.aws-route53-targets" = "^1.18.0"
"aws-cdk.aws-s3" = "^1.18.0"
"aws-cdk.aws-s3-assets" = "^1.18.0"
"aws-cdk.aws-secretsmanager" = "^1.18.0"
"aws-cdk.aws-servicediscovery" = "^1.18.0"
"aws-cdk.aws-sns" = "^1.18.0"
"aws-cdk.aws-sns-subscriptions" = "^1.18.0"
"aws-cdk.aws-sqs" = "^1.18.0"
"aws-cdk.aws-ssm" = "^1.18.0"
"aws-cdk.aws-stepfunctions" = "^1.18.0"
"aws-cdk.core" = "^1.18.0"
"aws-cdk.cx-api" = "^1.18.0"
"aws-cdk.region-info" = "^1.18.0"
aws-lambda = "^1.0.0"
boto3 = "^1.10.39"
botocore = "^1.13.39"
cattrs = "^0.9.0"
cfn-flip = "^1.2.2"
Click = "^7.0"
docutils = "^0.15.2"
jmespath = "^0.9.4"
jsii = "^0.20.11"
publication = "^0.0.3"
PyYAML = "^5.2"
s3transfer = "^0.2.1"
six = "^1.13.0"
troposphere = "^2.5.3"
urllib3 = "^1.25.7"
typing-extensions = "^3.7.4"
"aws-cdk.aws_lambda_event_sources" = "^1.19.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
